STLSAFE-master/STLSAFE-master/requirements.txt
numpy
torch==2.0.1+cpu
torchvision==0.15.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit
pillow
matplotlib
gdown

